requirements.txt
asn1crypto==1.4.0
attrs==20.3.0
azure-common==1.1.25
azure-core==1.8.2
azure-storage-blob==12.5.0
Babel==2.9.0
base58==2.0.1
bitarray==1.2.2
boto3==1.15.18
botocore==1.18.18
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.3.2
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
Flask==1.1.2
Flask-Babel==2.0.0
Flask-HTTPAuth==4.1.0
Flask-Table==0.5.0
Flask-WTF==0.14.3
gunicorn==20.0.4
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.7.0a1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
lru-dict==1.1.6
MarkupSafe==1.1.1
msrest==0.6.19
multiaddr==0.0.9
netaddr==0.8.0
numpy==1.19.4
oauthlib==3.1.0
oscrypto==1.2.1
pandas==1.0.5
parsimonious==0.8.1
plotly==4.12.0
protobuf==3.13.0
pyarrow==0.17.1
pycparser==2.20
pycryptodome==3.9.9
pycryptodomex==3.9.9
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rlp==2.0.0
s3transfer==0.3.3
six==1.15.0
snowflake-connector-python==2.3.5
toolz==0.11.1
urllib3==1.25.11
varint==1.0.2
web3==5.13.0
websockets==8.1.0
Werkzeug==1.0.1
WTForms==2.3.3
